Fri, Nov 1, 2024, 7:53 PM +00:00

+++ test2.c
@@ -4,6 +4,6 @@ int main()
 {
     int x = 2;
     int y = 3;
-    int z = x / 3;
-    printf("%d\n", z);
+    float z = x / 3;
+    printf("%f\n", z);
 }
